requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
Pillow>=9.5.0
librosa>=0.10.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
optuna>=3.0.0
tensorboard>=2.12.0
torchviz>=0.0.2
tqdm>=4.65.0